+++ js/constants.js
@@ -1,0 +1,30 @@
+const DESCRIPTION_LENDHT = 25; //массив из 25 объектов
+const DESCRIPTION = [
+  'разработка ',
+  'закат',
+  'котики',
+  'путешествие',
+  'природа',
+  'кофе',
+  'завтрак',
+  'свадьба',
+  'пляж',
+  'море',
+  'солнце',
+  'машины',
+  'ресторан',
+  'фитнес зал',
+  'животные',
+  'знакомства',
+  'девайсы',
+  'ремонт',
+  'квартиры',
+  'английский',
+  'обучение',
+  'продажи',
+  'индивидуальные предприниматели',
+  'самозанятые',
+  'курс валют'
+]; // создаем сами объекты массива
+
+export {DESCRIPTION_LENDHT, DESCRIPTION};
